Moved about and list commands to new Help.cs command file
