+++ LeafBot/LeafBot/Program.cs
@@ -152,6 +152,7 @@ namespace LeafBot
       Commands.RegisterCommands<Games>();
       Commands.RegisterCommands<Searches>();
       Commands.RegisterCommands<Counters>();
+      Commands.RegisterCommands<Help>();
 
       // Set custom helpformatter
       Commands.SetHelpFormatter<HelpFormatter>();
